+++ apps/omsorgspengerutbetaling-soknad/e2e/cypress/e2e/contextConfig.ts
@@ -1,6 +1,7 @@
 import { cyApiMockData } from './data/cyApiMockData';
-const PUBLIC_PATH = '/soknad';
-const API = 'http://localhost:8089';
+import { PUBLIC_PATH } from './utils';
+
+const INTERCEPT_PATH = `*/**`;
 
 const getUrlForStep = (step?) => {
     const url = `${PUBLIC_PATH}/soknad${step ? `/${step}` : '/velkommen'}`;
@@ -16,17 +17,27 @@ export const contextConfig = (props?: ConfigProps) => {
     const { mellomlagring, step } = props || {};
 
     beforeEach('intercept mellomlagring og levere tomt objekt', () => {
-        cy.intercept(`GET`, `${API}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`, mellomlagring || { noData: 1 });
-        cy.intercept(`DELETE`, `${API}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`, mellomlagring || {});
-        cy.intercept(`PUT`, `${API}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`, {}).as('putMellomlagring');
-        cy.intercept(`POST`, `${API}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`, {});
-        cy.intercept(`POST`, `${API}/omsorgspenger-utbetaling-snf/innsending`, {}).as('innsending');
-        cy.intercept('POST', `${API}/vedlegg`, {
+        cy.intercept(
+            `GET`,
+            `${INTERCEPT_PATH}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`,
+            mellomlagring || { noData: 1 },
+        );
+        cy.intercept(`DELETE`, `${INTERCEPT_PATH}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`, mellomlagring || {});
+        cy.intercept(`PUT`, `${INTERCEPT_PATH}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`, {}).as('putMellomlagring');
+        cy.intercept(`POST`, `${INTERCEPT_PATH}/mellomlagring/OMSORGSPENGER_UTBETALING_SNF`, {});
+        cy.intercept(`POST`, `${INTERCEPT_PATH}/omsorgspenger-utbetaling-snf/innsending`, {}).as('innsending');
+        cy.intercept('POST', `${INTERCEPT_PATH}/vedlegg`, {
             location: '/vedlegg',
             headers: { Location: '/vedlegg', 'access-control-expose-headers': 'Location' },
         });
-        cy.intercept('GET', `${API}/oppslag/soker*`, cyApiMockData.søkerMock).as('getSøker');
-        cy.intercept('GET', `${API}/oppslag/barn*`, { barn: props?.barn } || cyApiMockData.barnMock).as('getBarn');
+        cy.intercept('POST', `/vedlegg`, {
+            location: '/vedlegg',
+            headers: { Location: '/vedlegg', 'access-control-expose-headers': 'Location' },
+        });
+        cy.intercept('GET', `${INTERCEPT_PATH}/oppslag/soker*`, cyApiMockData.søkerMock).as('getSøker');
+        cy.intercept('GET', `${INTERCEPT_PATH}/oppslag/barn*`, { barn: props?.barn } || cyApiMockData.barnMock).as(
+            'getBarn',
+        );
         cy.intercept('*.api.sanity.io', {});
     });