+++ apps/omsorgspengesoknad/src/app/søknad/steps/delt-bosted/DeltBostedForm.tsx
@@ -1,6 +1,4 @@
 import { Alert, Link } from '@navikt/ds-react';
-import React from 'react';
-import { FormattedMessage, useIntl } from 'react-intl';
 import Block from '@navikt/sif-common-core-ds/lib/atoms/block/Block';
 import FormBlock from '@navikt/sif-common-core-ds/lib/atoms/form-block/FormBlock';
 import FileUploadErrors from '@navikt/sif-common-core-ds/lib/components/file-upload-errors/FileUploadErrors';
@@ -8,20 +6,22 @@ import PictureScanningGuide from '@navikt/sif-common-core-ds/lib/components/pict
 import SifGuidePanel from '@navikt/sif-common-core-ds/lib/components/sif-guide-panel/SifGuidePanel';
 import { Attachment } from '@navikt/sif-common-core-ds/lib/types/Attachment';
 import {
+    MAX_TOTAL_ATTACHMENT_SIZE_BYTES,
     attachmentHasBeenUploaded,
     getTotalSizeOfAttachments,
-    MAX_TOTAL_ATTACHMENT_SIZE_BYTES,
 } from '@navikt/sif-common-core-ds/lib/utils/attachmentUtils';
 import intlHelper from '@navikt/sif-common-core-ds/lib/utils/intlUtils';
-import { getTypedFormComponents, ValidationError, ValidationResult } from '@navikt/sif-common-formik-ds/lib';
+import { ValidationError, ValidationResult, getTypedFormComponents } from '@navikt/sif-common-formik-ds/lib';
 import { getListValidator } from '@navikt/sif-common-formik-ds/lib/validation';
 import getIntlFormErrorHandler from '@navikt/sif-common-formik-ds/lib/validation/intlFormErrorHandler';
 import { validateAll } from '@navikt/sif-common-formik-ds/lib/validation/validationUtils';
+import React from 'react';
+import { FormattedMessage, useIntl } from 'react-intl';
 import { ApiEndpoint } from '../../../api/api';
 import FormikFileUploader from '../../../components/formik-file-uploader/FormikFileUploader';
 import { getUploadedAttachments } from '../../../utils/attachmentUtils';
 import { relocateToLoginPage } from '../../../utils/navigationUtils';
-import { validateAttachments, ValidateAttachmentsErrors } from '../../../utils/validateAttachments';
+import { ValidateAttachmentsErrors, validateAttachments } from '../../../utils/validateAttachments';
 import DeltBostedAvtaleAttachmentList from './DeltBostedAvtaleAttachmentList';
 
 interface Props {
@@ -75,7 +75,7 @@ const DeltBostedForm: React.FunctionComponent<Props> = ({ values, goBack, andreV
             onBack={goBack}>
             <Block padBottom="xl">
                 <SifGuidePanel>
-                    <p style={{ marginTop: 0 }}>
+                    <p>
                         <FormattedMessage id={'steg.deltBosted.helperTextPanel.1'} />
                     </p>
                     <p>