+++ apps/pleiepenger-i-livets-sluttfase-soknad/e2e/cypress/e2e/utils/medlemskap.ts
@@ -2,14 +2,13 @@ import * as dayjs from 'dayjs';
 import * as locale from 'dayjs/locale/nb';
 import * as isoWeek from 'dayjs/plugin/isoWeek';
 import {
-    getElement,
     getInputByName,
     getTestElement,
     selectRadioYes,
     submitModal,
     submitSkjema,
     selectRadioYesOrNo,
-} from '../utils';
+} from './cyHelpers';
 
 dayjs.extend(isoWeek);
 dayjs.locale(locale);
@@ -34,22 +33,22 @@ const fyllUtMedlemskapKomplett = () => {
     selectRadioYesOrNo('medlemskap-annetLandSiste12', true);
 
     getTestElement('bostedUtlandList-annetLandSiste12').within(() => {
-        getElement('button').contains('Legg til nytt utenlandsopphold').click();
+        cy.get('button').contains('Legg til nytt utenlandsopphold').click();
     });
     getInputByName('fom').click().type(fomTomMedlemskapSiste12).blur();
     getInputByName('tom').click().type(fomTomMedlemskapSiste12).blur();
-    getElement('select').select(2); // Valg land #2 fra listen
+    cy.get('select').select(2); // Valg land #2 fra listen
     submitModal();
 
     selectRadioYes('medlemskap-annetLandNeste12');
 
     getTestElement('bostedUtlandList-annetLandNeste12').within(() => {
-        getElement('button').contains('Legg til nytt utenlandsopphold').click();
+        cy.get('button').contains('Legg til nytt utenlandsopphold').click();
     });
 
     getInputByName('fom').click().type(fomTomMedlemskapNeste12).blur();
     getInputByName('tom').click().type(fomTomMedlemskapNeste12).blur();
-    getElement('select').select(2); // Valg land #2 fra listen
+    cy.get('select').select(2); // Valg land #2 fra listen
     submitModal();
 
     submitSkjema();
@@ -65,13 +64,13 @@ const oppsummeringTestMedlemskapEnkelt = () => {
 const oppsummeringTestMedlemskapKomplett = () => {
     getTestElement('oppsummering-medlemskap-utlandetSiste12').should((element) => expect('Ja').equal(element.text()));
     getTestElement('oppsummering-medlemskap-utlandetSiste12-list').within(() => {
-        getElement('li')
+        cy.get('li')
             .eq(0)
             .within(() => {
-                getElement('span')
+                cy.get('span')
                     .eq(0)
                     .should((element) => expect(expectedDateMedlemskapSiste12).equal(element.text()));
-                getElement('span')
+                cy.get('span')
                     .eq(1)
                     .should((element) => expect(expectedLand).equal(element.text()));
             });
@@ -79,13 +78,13 @@ const oppsummeringTestMedlemskapKomplett = () => {
 
     getTestElement('oppsummering-medlemskap-utlandetNeste12').should((element) => expect('Ja').equal(element.text()));
     getTestElement('oppsummering-medlemskap-utlandetNeste12-list').within(() => {
-        getElement('li')
+        cy.get('li')
             .eq(0)
             .within(() => {
-                getElement('span')
+                cy.get('span')
                     .eq(0)
                     .should((element) => expect(expectedDateMedlemskapNeste12).equal(element.text()));
-                getElement('span')
+                cy.get('span')
                     .eq(1)
                     .should((element) => expect(expectedLand).equal(element.text()));
             });