+++ apps/pleiepenger-i-livets-sluttfase-soknad/src/app/utils/getTilgjengeligEndringsperiode.ts
@@ -4,6 +4,6 @@ import dayjs from 'dayjs';
 export const getTilgjengeligEndringsperiode = (): DateRange => {
     return {
         from: dayjs().startOf('month').subtract(3, 'months').toDate(),
-        to: dayjs().add(1, 'month').toDate(),
+        to: dayjs().add(4, 'weeks').toDate(),
     };
 };