+++ apps/pleiepenger-i-livets-sluttfase-soknad/src/app/utils/søknadsdataToApiData/getFrilansApiDataFromSøknadsdata.ts
@@ -7,8 +7,9 @@ import { getArbeidIPeriodeApiDataFromSøknadsdata } from './getArbeidIPeriodeApi
 export const getFrilansApiDataFromSøknadsdata = (
     søknadsperiode: DateRange,
     dagerMedPleie: Date[],
+    skalJobbeIPerioden: boolean,
     frilans?: ArbeidFrilansSøknadsdata,
-    arbeidIperiode?: ArbeidIPeriodeSøknadsdata,
+    arbeidIPeriode?: ArbeidIPeriodeSøknadsdata,
 ): FrilansApiData | undefined => {
     if (!frilans) {
         return undefined;
@@ -18,7 +19,7 @@ export const getFrilansApiDataFromSøknadsdata = (
             return undefined;
 
         case 'sluttetISøknadsperiode':
-            if (arbeidIperiode) {
+            if (arbeidIPeriode) {
                 return {
                     harHattInntektSomFrilanser: true,
                     startdato: frilans.startdato,
@@ -27,7 +28,8 @@ export const getFrilansApiDataFromSøknadsdata = (
                     arbeidsforhold: {
                         jobberNormaltTimer: frilans.jobberNormaltTimer,
                         arbeidIPeriode: getArbeidIPeriodeApiDataFromSøknadsdata(
-                            arbeidIperiode,
+                            skalJobbeIPerioden,
+                            arbeidIPeriode,
                             søknadsperiode,
                             frilans.jobberNormaltTimer,
                             dagerMedPleie,
@@ -37,7 +39,7 @@ export const getFrilansApiDataFromSøknadsdata = (
             } else return undefined;
 
         case 'pågående':
-            if (arbeidIperiode) {
+            if (arbeidIPeriode) {
                 return {
                     harHattInntektSomFrilanser: true,
                     startdato: frilans.startdato,
@@ -45,7 +47,8 @@ export const getFrilansApiDataFromSøknadsdata = (
                     arbeidsforhold: {
                         jobberNormaltTimer: frilans.jobberNormaltTimer,
                         arbeidIPeriode: getArbeidIPeriodeApiDataFromSøknadsdata(
-                            arbeidIperiode,
+                            skalJobbeIPerioden,
+                            arbeidIPeriode,
                             søknadsperiode,
                             frilans.jobberNormaltTimer,
                             dagerMedPleie,