+++ packages/sif-common-core-ds/src/i18n/common.nb.ts
@@ -1,0 +1,38 @@
+export const commonMessagesNb = {
+    Ja: 'Ja',
+    Nei: 'Nei',
+    VetIkke: 'Jeg er usikker',
+
+    'hjelpetekst.skjul': 'Lukk hjelpetekst',
+    'hjelpetekst.vis': 'Vis hjelpetekst',
+    'backlink.label': 'Tilbake',
+
+    'locale.nb': 'Bokmål - norsk',
+    'locale.nn': 'Nynorsk - norsk',
+    'locale.en': 'Nynorsk - norsk',
+    'locale.no-NB': 'Bokmål - norsk',
+    'locale.no-NN': 'Nynorsk - norsk',
+
+    'page.defaultMainRoleLabel': 'Hovedinnhold',
+
+    'formikValidationErrorSummary.tittel': 'Du må rette opp i følgende feil:',
+    'formikForm.submitButton.label': 'Ok',
+    'formikForm.cancelButton.label': 'Avbryt',
+
+    timer: '{timer, plural, one {# time} other {# timer}}',
+    minutter: '{minutter, plural, one {# minutt} other {# minutter}}',
+    timerOgMinutter:
+        '{timer, plural, one {# time} other {# timer}} og {minutter, plural, one {# minutt} other {# minutter}}',
+    dager: '{dager, plural, one {# dag} other {# dager}}',
+
+    'vedleggsliste.fjernKnapp': 'Fjern',
+    ariaStepInfo: 'Steg {currentStep} av {steps}',
+
+    'dataRangeValidation.form.validation.required': 'Feltet er påkrevd',
+    'dataRangeValidation.form.validation.dateOutsideRange':
+        'Første gyldige dato er {fom}, og siste gyldige dato er {tom}',
+    'dataRangeValidation.form.validation.dateBeforeMinDate': 'Første gyldige dato er {fom}',
+    'dataRangeValidation.form.validation.dateAfterMaxDate': 'Siste gyldige dato er {tom}',
+    'dataRangeValidation.form.validation.fromDateAfterToDate': 'Fra-dato må være lik eller før til-dato',
+    'dataRangeValidation.form.validation.toDateBeforeFromDate': 'Til-dato må være lik eller etter fra-dato',
+};