+++ packages/sif-common-ui/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @navikt/sif-common-ui
 
+## 0.0.10
+
+### Patch Changes
+
+-   Pakkeoppdatering som fikser datepicker + modal bug
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@1.1.2
+    -   @navikt/sif-common-core-ds@9.8.6
+    -   @navikt/sif-common-utils@3.40.7
+
+## 0.0.9
+
+### Patch Changes
+
+-   71381778: Patch pakkeoppdateringer
+-   Updated dependencies [71381778]
+    -   @navikt/sif-common-formik-ds@1.1.1
+    -   @navikt/sif-common-core-ds@9.8.5
+    -   @navikt/sif-common-utils@3.40.6
+
+## 0.0.8
+
+### Patch Changes
+
+-   60eb7ee9: Pakkeoppdateringer. Endret import av DatePickerProps i sif-common-formik-ds
+-   Updated dependencies [60eb7ee9]
+    -   @navikt/sif-common-formik-ds@1.1.0
+    -   @navikt/sif-common-core-ds@9.8.4
+    -   @navikt/sif-common-utils@3.40.5
+
+## 0.0.7
+
+### Patch Changes
+
+-   2a1e2526: Pakkeoppdateringer
+-   Updated dependencies [2a1e2526]
+    -   @navikt/sif-common-formik-ds@1.0.5
+    -   @navikt/sif-common-core-ds@9.8.3
+    -   @navikt/sif-common-utils@3.40.4
+
+## 0.0.6
+
+### Patch Changes
+
+-   Pakkeoppdateringer
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@1.0.4
+    -   @navikt/sif-common-core-ds@9.8.2
+    -   @navikt/sif-common-utils@3.40.3
+
+## 0.0.5
+
+### Patch Changes
+
+-   81e7e155: Bruke BodyLong i alle SifGuidePanel
+-   Updated dependencies [81e7e155]
+    -   @navikt/sif-common-formik-ds@1.0.3
+    -   @navikt/sif-common-core-ds@9.8.1
+    -   @navikt/sif-common-utils@3.40.2
+
 ## 0.0.4
 
 ### Patch Changes